Update seen messages goes through backend first
I realized that I wanted to actually modify the backend before
continuing to change the store (I had created the code for changing the
store). So I refactored the code, adding the function
`updateSeenMessage` in `thunkCreators.js` which is the function that
sends the post request to the back-end.

So now, when a user sees unread messages, the sequence of operations is
this: dispatch(updateSeenMessage) --> *updates database* -->
dispatch(setSeenAll) --> *updates store*